move: error code 경로 수정
