+++ src/main/java/SeCause/SeCause_be/domain/analysis/exception/AnalysisException.java
@@ -1,6 +1,6 @@
 package SeCause.SeCause_be.domain.analysis.exception;
 
-import SeCause.SeCause_be.domain.analysis.code.AnalysisErrorCode;
+import SeCause.SeCause_be.domain.analysis.exception.code.AnalysisErrorCode;
 import SeCause.SeCause_be.global.apiPayload.exception.GeneralException;
 
 public class AnalysisException extends GeneralException {
