+++ src/main/java/SeCause/SeCause_be/domain/analysis/exception/code/AnalysisErrorCode.java
@@ -1,4 +1,4 @@
-package SeCause.SeCause_be.domain.analysis.code;
+package SeCause.SeCause_be.domain.analysis.exception.code;
 
 import SeCause.SeCause_be.global.apiPayload.code.BaseErrorCode;
 import lombok.AllArgsConstructor;
