+++ src/main/java/SeCause/SeCause_be/domain/projectRepository/exception/ProjectRepositoryException.java
@@ -1,6 +1,6 @@
 package SeCause.SeCause_be.domain.projectRepository.exception;
 
-import SeCause.SeCause_be.domain.projectRepository.code.ProjectRepositoryErrorCode;
+import SeCause.SeCause_be.domain.projectRepository.exception.code.ProjectRepositoryErrorCode;
 import SeCause.SeCause_be.global.apiPayload.exception.GeneralException;
 
 public class ProjectRepositoryException extends GeneralException {
