+++ src/main/java/SeCause/SeCause_be/domain/projectRepository/service/RepositoryIssueService.java
@@ -1,6 +1,6 @@
 package SeCause.SeCause_be.domain.projectRepository.service;
 
-import SeCause.SeCause_be.domain.analysis.code.AnalysisErrorCode;
+import SeCause.SeCause_be.domain.analysis.exception.code.AnalysisErrorCode;
 import SeCause.SeCause_be.domain.analysis.exception.AnalysisException;
 import SeCause.SeCause_be.domain.analysis.repository.AnalysisResultRepository;
 import SeCause.SeCause_be.domain.projectRepository.dto.RepositoryIssueDetailResponse;
