+++ src/main/java/SeCause/SeCause_be/domain/projectRepository/validator/ProjectRepositoryValidator.java
@@ -1,6 +1,6 @@
 package SeCause.SeCause_be.domain.projectRepository.validator;
 
-import SeCause.SeCause_be.domain.projectRepository.code.ProjectRepositoryErrorCode;
+import SeCause.SeCause_be.domain.projectRepository.exception.code.ProjectRepositoryErrorCode;
 import SeCause.SeCause_be.domain.projectRepository.exception.ProjectRepositoryException;
 import SeCause.SeCause_be.domain.projectRepository.repository.ProjectRepositoryRepository;
 import lombok.RequiredArgsConstructor;
